third_party/blink/web_tests/external/wpt/preload/preload-time-to-fetch.https-expected.txt

This is a testharness.js-based test.
[PASS] Changing a preload href should trigger a fetch
[PASS] Changing a preload "as" from a previously non-matching destination should trigger a fetch
[PASS] Changing a preload "type" (non-matching->matching) should trigger a fetch
[PASS] Removing a preload non-matching "type" should trigger a fetch
[FAIL] Removing a preload matching "type" should not trigger a fetch
  assert_array_equals: expected property 1 to be "timeout" but got "load" (expected array ["load", "timeout"] got ["load", "load"])
[PASS] Changing a preload media attribute (non matching->matching) should trigger a fetch
[PASS] Changing a preload media attribute (non matching->non matching) should not trigger a fetch
[PASS] Changing a preload media attribute (matching->matching) should not trigger a new fetch
[PASS] Removing a matching preload media attribute should not trigger a new fetch
[PASS] Removing a non-matching preload media attribute should trigger a new fetch
Harness: the test ran to completion.
